+++ router.php
@@ -25,7 +25,7 @@
   // do some URL handling
   $result = false;
   $path   = parse_url($_SERVER["REQUEST_URI"], PHP_URL_PATH);
-  if (preg_match_array(["@^\/contact$@", "@^\/newsletter$@", "@^\/register$@", "@^\/subscribed$@",
+  if (preg_match_array(["@^\/contact$@", "@^\/newsletter$@", "@^\/register$@", "@^\/send$@", "@^\/subscribed$@",
                         "@^\/verified$@", "@^\/verify$@"], $path)) {
     // pretty URLs
     require_once(__DIR__."$path.php");